requirements.txt
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.11.1
PyYAML~=6.0